requirements.txt
Jinja2==3.1.2
Markdown==3.5
MarkupSafe==2.1.3
Pygments==2.16.1
PyYAML==6.0.1
